requirements.txt
Django==1.9.2
Pillow==3.1.1
python-social-auth==0.2.14
djangorestframework==3.3.2
django-filter==0.12.0
